requirements.txt
mysql-connector-python>=9.0.0
streamlit>=1.0.0
pandas>=2.0.0
argon2-cffi>=23.1.0
